Move code to generic output
